Move lrucache to internal, remove max default value
